Pass connector_id to /auth on completion of device auth flow
This allows users to select a connector when initiating device auth.
They simply need to append "connector_id" to the query string of
"/device"  alongside "user_code". The form will submit the chosen
connector to Dex which will add the connector to  "/auth" to skip the
webpage for selecting a connector.